+++ src/views/add-boiler-house-view/layout.tsx
@@ -1,73 +1,56 @@
-// import React, { useState } from "react";
-// import { Link as RouterLink } from "react-router-dom";
+import React, { useState } from "react";
 
-// import { NewAsset } from "../../components/new-asset-form";
-// import { locale } from "../../services";
+import { ErrorSummary, Link, StatusBox } from "@mtfh/common/lib/components";
+import { Link as RouterLink, useParams } from "react-router-dom";
+import { Asset, useAsset } from "@mtfh/common/lib/api/asset/v1";
+import { AddBoilerHouseForm } from "../../components/add-boiler-house-form";
 
-// import { CreateNewAssetRequest } from "@mtfh/common/lib/api/asset/v1";
-// import { ErrorSummary, Link, StatusBox } from "@mtfh/common/lib/components";
+interface Props {
+  assetId: string;
+  asset: Asset;
+}
 
-// export const NewPropertyLayout = (): JSX.Element => {
-//   const [newProperty, setNewProperty] = useState<CreateNewAssetRequest | null>();
-//   const [showSuccess, setShowSuccess] = useState<boolean>(false);
-//   const [showError, setShowError] = useState<boolean>(false);
-//   const [errorHeading, setErrorHeading] = useState<string | null>(null);
-//   const [errorDescription, setErrorDescription] = useState<string | null>(null);
+export const AddBoilerHouseLayout = ({ assetId, asset }: Props) => {
+  const [showSuccess, setShowSuccess] = useState(false);
+  const [requestError, setRequestError] = useState<string | null>(null);
 
-//   const renderNewPropertyLink = () => {
-//     return (
-//       // Only include link if visible on MMH
-//       (newProperty?.assetType === "Dwelling" ||
-//         newProperty?.assetType === "LettableNonDwelling") && (
-//         <div>
-//           {" "}
-//           <Link as={RouterLink} to={`/property/${newProperty?.id}`}>
-//             View property
-//           </Link>{" "}
-//         </div>
-//       )
-//     );
-//   };
+  return (
+    <>
+      <Link as={RouterLink} to={`/property/${assetId}`} variant="back-link">
+        Back to asset
+      </Link>
 
-//   return (
-//     <>
-//       <Link as={RouterLink} to="#" variant="back-link">
-//         Back
-//       </Link>
-//       <h1 className="lbh-heading-h1">New property</h1>
-//       <span
-//         id="new-property-disclaimer"
-//         className="govuk-hint lbh-hint"
-//         data-testid="new-property-disclaimer"
-//       >
-//         A new property will immediately be available for users of Manage My Home.
-//         Information is validated in most fields, but all protections are not in place, so
-//         some selections may result in mismatched information. It is recommended to
-//         thoroughly check the submission before adding the new asset.{" "}
-//       </span>
-//       {showSuccess && (
-//         <StatusBox variant="success" title={locale.assets.newPropertyAddedSuccessMessage}>
-//           {renderNewPropertyLink()}
-//         </StatusBox>
-//       )}
+      {showSuccess && (
+        <StatusBox
+          variant="success"
+          title={
+            <span>
+              The boiler house has been added successfully.{" "}
+              <Link as={RouterLink} to={`/property/${assetId}`}>
+                Return to property
+              </Link>
+            </span>
+          }
+        />
+      )}
 
-//       {showError && (
-//         <ErrorSummary
-//           id="patch-asset-error"
-//           title={errorHeading || ""}
-//           description={errorDescription || undefined}
-//         />
-//       )}
+      <h1 className="lbh-heading-h1">Add boiler house to property</h1>
+      <span className="govuk-caption-m lbh-caption">Some subtitle text</span>
 
-//       <section>
-//         <NewAsset
-//           setShowError={setShowError}
-//           setErrorHeading={setErrorHeading}
-//           setErrorDescription={setErrorDescription}
-//           setShowSuccess={setShowSuccess}
-//           setNewProperty={setNewProperty}
-//         />
-//       </section>
-//     </>
-//   );
-// };
+      {requestError && (
+        <ErrorSummary
+          id="patch-asset-error"
+          title="Unexpected error"
+          description={requestError || undefined}
+        />
+      )}
+
+      <AddBoilerHouseForm
+        setShowSuccess={setShowSuccess}
+        setRequestError={setRequestError}
+        asset={asset}
+        assetId={assetId}
+      />
+    </>
+  );
+};
